Added Tutorial Script
First you go like this, spin around, and STOP. Double-take three times.
One, two, three. And...calva-thrust! Whoo! Whoo! Step on your right foot, don't forget it!
Now it's time to bring it around town, bring it around town.
Then you do this, then this, and this, then this, then that, then this and that, and then...